Extract interface for BindingFilterContinuation
