+++ modules/core/arrow-test/src/main/kotlin/arrow/test/laws/MonadFilterLaws.kt
@@ -1,14 +1,13 @@
 package arrow.test.laws
 
-import arrow.typeclasses.Eq
 import arrow.Kind
 import arrow.mtl.MonadFilter
-import arrow.mtl.bindingFilter
 import arrow.mtl.monadFilter
 import arrow.syntax.monad.flatMap
 import arrow.test.generators.genApplicative
 import arrow.test.generators.genConstructor
 import arrow.test.generators.genFunctionAToB
+import arrow.typeclasses.Eq
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
